+++ packages/strapi-plugin-content-manager/test/api/basic.test.e2e.js
@@ -1,0 +1,125 @@
+const { registerAndLogin } = require('../../../../test/helpers/auth');
+const createModelsUtils = require('../../../../test/helpers/models');
+const { createAuthRequest } = require('../../../../test/helpers/request');
+
+let rq;
+let modelsUtils;
+let data = {
+  products: [],
+  productsWithDP: [],
+  productsWithCompo: [],
+  productsWithCompoAndDP: [],
+};
+
+const product = {
+  attributes: {
+    name: {
+      type: 'string',
+    },
+    description: {
+      type: 'text',
+    },
+  },
+  connection: 'default',
+  name: 'product',
+  description: '',
+  collectionName: '',
+};
+
+const compo = {
+  name: 'compo',
+  attributes: {
+    name: {
+      type: 'string',
+      required: true,
+    },
+    description: {
+      type: 'text',
+      minLength: 4,
+      maxLength: 30,
+    },
+  },
+};
+
+describe('CM API - Basic', () => {
+  beforeAll(async () => {
+    const token = await registerAndLogin();
+    rq = createAuthRequest(token);
+
+    modelsUtils = createModelsUtils({ rq });
+    await modelsUtils.createComponent(compo);
+    await modelsUtils.createContentTypes([product]);
+  }, 60000);
+
+  afterAll(async () => {
+    await modelsUtils.deleteContentTypes(['product']);
+    await modelsUtils.deleteComponent('default.compo');
+  }, 60000);
+
+  test('Create Products', async () => {
+    const product = {
+      name: 'Product 1',
+      description: 'Product description',
+    };
+    const res = await rq({
+      method: 'POST',
+      url: '/content-manager/explorer/application::product.product',
+      body: product,
+    });
+
+    expect(res.statusCode).toBe(200);
+    expect(res.body).toMatchObject(product);
+    expect(res.body.published_at).toBeUndefined();
+    data.products.push(res.body);
+  });
+
+  test('Read Products', async () => {
+    const res = await rq({
+      method: 'GET',
+      url: '/content-manager/explorer/application::product.product',
+    });
+
+    expect(res.statusCode).toBe(200);
+    expect(Array.isArray(res.body)).toBe(true);
+    expect(res.body).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          name: 'Product 1',
+          description: 'Product description',
+        }),
+      ])
+    );
+    res.body.forEach(p => expect(p.published_at).toBeUndefined());
+  });
+
+  test('Update Products', async () => {
+    const product = {
+      name: 'Product 1 updated',
+      description: 'Updated Product description',
+    };
+    const res = await rq({
+      method: 'PUT',
+      url: `/content-manager/explorer/application::product.product/${data.products[0].id}`,
+      body: product,
+    });
+
+    expect(res.statusCode).toBe(200);
+    expect(res.body).toMatchObject(product);
+    expect(res.body.id).toEqual(data.products[0].id);
+    expect(res.body.published_at).toBeUndefined();
+    data.products[0] = res.body;
+  });
+
+  test('Delete Products', async () => {
+    const res = await rq({
+      method: 'DELETE',
+      url: `/content-manager/explorer/application::product.product/${data.products[0].id}`,
+    });
+
+    expect(res.statusCode).toBe(200);
+    expect(res.body).toMatchObject(data.products[0]);
+    expect(res.body.id).toEqual(data.products[0].id);
+    expect(res.body.published_at).toBeUndefined();
+    data.products.shift();
+  });
+});
