+++ packages/strapi-plugin-content-manager/test/fields/boolean.test.e2e.js
@@ -58,8 +58,6 @@ describe('Test type boolean', () => {
       body: { field: 1 },
     });
 
-    console.log('res.body', JSON.stringify(res.body, null, 2));
-
     expect(res.statusCode).toBe(200);
     expect(res.body).toMatchObject({
       field: true,
